Rename MetadataSample to IntermediateFormatSample and move to different namespace
